# swagger.yaml
openapi: 3.0.0
info:
  title: _TODO
  description: Transfer your personal data from multiple data providers to your Solid Pod.
  version: _TODO
servers:
  - url: http://localhost:4000/
    description: Optional server description, e.g. Main (production) server

paths:
  /api-docs:
    get:
      summary: Gets these docs
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
  /{provider}/callback:
    get:
      summary: Updates the provider tokens and redirects back to root
      responses:
        "302":
          description: OK
    parameters:
      - name: provider
        in: path
        description: The provider
        required: true
        schema:
          $ref: "#/components/schemas/ProviderModel"
  /rmlmapper:
    post:
      summary: Execute the RMLMapper
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: 
                  $ref: "#/components/schemas/ProviderModel"
                file:
                  type: string
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: object
                properties:
                  rdf:
                    type: string
                  prov:
                    type: object
        401:
          description: No credentials
        500:
          description: Not OK
          content:
            text/plain:
              schema:
                type: string
  /status/{provider}/connected:
    get:
      summary: Returns whether the given provider is connected (i.e. whether the backend contains tokens for that provider.
      parameters:
        - name: provider
          in: path
          description: Data provider (e.g. flickr)
          required: true
          schema:
            $ref: "#/components/schemas/ProviderModel"
      responses:
        200:
          description: Returns true or false, depending on whether the given provider is connected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    $ref: "#/components/schemas/ProviderModel"
                  connected:
                    type: boolean
  /configuration/{provider}/{configKey}:
    get:
      summary: Returns the configuration of a provider
      parameters:
        - name: provider
          in: path
          description: Data provider (e.g. flickr)
          required: true
          schema:
            $ref: "#/components/schemas/ProviderModel"
        - name: configKey
          in: path
          description: configuration key
          required: true
          schema:
            $ref: "#/components/schemas/ConfigurationKey"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
  /logout:
    post:
      summary: Log out. This will destroy the session.
      responses:
        200:
          description: Succesfully logged out from all connected providers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        500:
          description: Internal Server Error. Something went wrong when logging out.
components:
  schemas:
    ProviderModel:
      type: string
      enum:
        - imgur
        - flickr
        - google
    ConfigurationKey:
      type: string
      enum:
        - connect